CommentにdeleteCommentメソッドを移行
